Optimisations
cache transparent state in material
moved shadow shaders to shader manager
cache transform matrices
uint32 handles again

+++ source/engine/model_manager.h
@@ -15,8 +15,8 @@ namespace Engine
 
 	struct ModelHandle
 	{
-		uint16_t m_index = -1;
-		static ModelHandle Invalid() { return { (uint16_t)-1 }; };
+		uint32_t m_index = -1;
+		static ModelHandle Invalid() { return { (uint32_t)-1 }; };
 	};
 
 	class ModelManager
